requirements.txt
bcrypt==3.1.0
boto==2.42.0
cffi==1.7.0
Django==1.9.2
django-registration-redux==1.4
django-storages==1.4.1
Pillow==3.3.0
psycopg2==2.6.2
pycparser==2.14
six==1.10.0
